backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Bcrypt
Flask-CORS
PyMySQL
python-dotenv
pandas
scikit-learn